+++ graphics/cube.cpp
@@ -1,4 +1,4 @@
-#include "cube.h"
+#include "../include/cube.h"
 
 Cube::Cube(glm::vec3 cubeMin, glm::vec3 cubeMax) {
 	// Model matrix - Maybe decouple from cube?
@@ -145,7 +145,7 @@ void Cube::draw(const glm::mat4& viewProjMtx, Shader* shader) {
     glBindVertexArray(VAO);
 
     // Draw the points using triangles, indexed with the EBO
-    glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+    glDrawElements(GL_TRIANGLES, (int)indices.size(), GL_UNSIGNED_INT, 0);
 
     // Unbind the VAO and shader program
     glBindVertexArray(0);
